+++ test/PortalSpec.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import ReactTestUtils from 'react-addons-test-utils';
+import ReactTestUtils from 'react-dom/test-utils';
+
 import Portal from '../src/Portal';
 
 describe('Portal', function () {
@@ -10,12 +11,18 @@ describe('Portal', function () {
     render() {
       return (
         <div>
-          <Portal ref='p' {...this.props}>{this.props.overlay}</Portal>
+          <Portal
+            ref={(c) => { this.portal = c; }}
+            {...this.props}
+          >
+            {this.props.overlay}
+          </Portal>
         </div>
       );
     }
+
     getOverlayDOMNode = () => {
-      return this.refs.p.getOverlayDOMNode();
+      return this.portal.getOverlayDOMNode();
     }
   }
 
@@ -52,7 +59,13 @@ describe('Portal', function () {
   it('Should not render a null overlay', function() {
     class Container extends React.Component {
       render() {
-        return <Overlay ref='overlay' container={this} overlay={null} />;
+        return (
+          <Overlay
+            ref={(c) => { this.overlay = c; }}
+            container={this}
+            overlay={null}
+          />
+        );
       }
     }
 
@@ -60,13 +73,20 @@ describe('Portal', function () {
       <Container />
     );
 
-    assert.equal(instance.refs.overlay.getOverlayDOMNode(), null);
+    assert.equal(instance.overlay.getOverlayDOMNode(), null);
   });
 
   it('Should render only an overlay', function() {
     class OnlyOverlay extends React.Component {
       render() {
-        return <Portal ref='p' {...this.props}>{this.props.overlay}</Portal>;
+        return (
+          <Portal
+            ref={(c) => { this.portal = c; }}
+            {...this.props}
+          >
+            {this.props.overlay}
+          </Portal>
+        );
       }
     }
 
@@ -74,7 +94,7 @@ describe('Portal', function () {
       <OnlyOverlay overlay={<div id="test1" />} />
     );
 
-    assert.equal(overlayInstance.refs.p.getOverlayDOMNode().nodeName, 'DIV');
+    assert.equal(overlayInstance.portal.getOverlayDOMNode().nodeName, 'DIV');
   });
 
   it('Should change container on prop change', function() {
@@ -84,8 +104,12 @@ describe('Portal', function () {
       render() {
         return (
           <div>
-            <div ref='d' />
-            <Portal ref='p' {...this.props} container={this.state.container}>
+            <div ref={(c) => { this.container = c; }} />
+            <Portal
+              ref={(c) => { this.portal = c; }}
+              {...this.props}
+              container={this.state.container}
+            >
               {this.props.overlay}
             </Portal>
           </div>
@@ -97,15 +121,16 @@ describe('Portal', function () {
       <ContainerTest overlay={<div id="test1" />} />
     );
 
-    assert.equal(overlayInstance.refs.p._portalContainerNode.nodeName, 'BODY');
-    overlayInstance.setState({container: overlayInstance.refs.d})
-    assert.equal(overlayInstance.refs.p._portalContainerNode.nodeName, 'DIV');
+    assert.equal(overlayInstance.portal._portalContainerNode.nodeName, 'BODY');
+    overlayInstance.setState({container: overlayInstance.container})
+    assert.equal(overlayInstance.portal._portalContainerNode.nodeName, 'DIV');
 
-    ReactDOM.unmountComponentAtNode(ReactDOM.findDOMNode(overlayInstance).parentNode);
+    ReactDOM.unmountComponentAtNode(
+      ReactDOM.findDOMNode(overlayInstance).parentNode,
+    );
   });
 
   it('Should unmount when parent unmounts', function() {
-
     class Parent extends React.Component {
       state = {show: true};
       render() {
@@ -121,8 +146,8 @@ describe('Portal', function () {
       render() {
         return (
           <div>
-            <div ref='d' />
-            <Portal ref='p' container={() => this.refs.d}>
+            <div ref={(c) => { this.container = c; }} />
+            <Portal container={() => this.container}>
               <div id="test1" />
             </Portal>
           </div>