+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## [v0.8.1]
+> 2017-09-13
+
+- **Bugfix:** Use `keydown` instead of incorrect `keyup` for `<Modal>` close keyboard event ([#195])
+
+[v0.8.1]: https://github.com/react-bootstrap/react-overlays/compare/v0.8.0...v0.8.1
+[#195]: https://github.com/react-bootstrap/react-overlays/pull/195
+
+
+## [v0.8.0]
+> 2017-07-03
+
+- **Feature:** Remove `<Transition>` and depend on react-transition-group@2.0.0 ([#184])
+
+[v0.8.0]: https://github.com/react-bootstrap/react-overlays/compare/v0.7.0...v0.8.0
+[#184]: https://github.com/react-bootstrap/react-overlays/pull/184
+
+
+## [v0.7.0]
+> 2017-04-22
+
+- **Chore:** Update depedencies to avoid React deprecation warnings
+- **Chore:** Use function refs ([#159])
+
+[v0.7.0]: https://github.com/react-bootstrap/react-overlays/compare/v0.6.12...v0.7.0
+[#159]: https://github.com/react-bootstrap/react-overlays/pull/159
+
+
 ## [v0.6.12]
 > 2017-03-06